Added WebSocketNonNegotiableExtensionBuilder
